[PL] extract calculation of coaxial thermal resistances to new h file.
